+++ packages/core/phase3/lib/hearingDetails/getConvictionDateFromPncAdjudicationIfOffenceIsAdjournedSineDie.test.ts
@@ -1,7 +1,7 @@
-import type { Offence } from "../../types/AnnotatedHearingOutcome"
+import type { Offence } from "../../../types/AnnotatedHearingOutcome"
 
-import generateAhoFromOffenceList from "../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
-import generateAhoMatchingPncAdjudicationAndDisposals from "../../phase2/tests/helpers/generateAhoMatchingPncAdjudicationAndDisposals"
+import generateAhoFromOffenceList from "../../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
+import generateAhoMatchingPncAdjudicationAndDisposals from "../../../phase2/tests/helpers/generateAhoMatchingPncAdjudicationAndDisposals"
 import getConvictionDateFromPncAdjudicationIfOffenceIsAdjournedSineDie from "./getConvictionDateFromPncAdjudicationIfOffenceIsAdjournedSineDie"
 
 describe("getConvictionDateFromPncAdjudicationIfOffenceIsAdjournedSineDie", () => {
